Move config file test over to utils test suite
